app/requirements/base.txt
Django==4.1
django-cors-headers==3.13.0
django-extensions==3.2.0
Pillow==9.2.0
psycopg2==2.9.3
pytest==7.1.3
pytest-django==4.5.2

-e git+ssh://git@github.com/deniskolokol/genery.git#egg=genery
